+++ src/onegov.user/onegov/user/__init__.py
@@ -5,8 +5,8 @@ log.addHandler(logging.NullHandler())  # noqa
 from translationstring import TranslationStringFactory
 _ = TranslationStringFactory('onegov.user')  # noqa
 
-from onegov.user.auth import Auth
+from onegov.user.auth import Auth, is_valid_yubikey
 from onegov.user.collection import UserCollection
 from onegov.user.model import User
 
-__all__ = ['Auth', 'User', 'UserCollection']
+__all__ = ['Auth', 'is_valid_yubikey', 'User', 'UserCollection']